Moved the location of some functions out of jsonreader to json downloader
