+++ src/libopenrave-core/jsonparser/jsondownloader.h
@@ -34,6 +34,79 @@
 
 
 
+namespace OpenRAVE
+{
+
+
+static bool _EndsWith(const std::string& fullString, const std::string& endString) {
+    if (fullString.length() >= endString.length()) {
+        return fullString.compare(fullString.length() - endString.length(), endString.length(), endString) == 0;
+    }
+    return false;
+}
+
+
+// Need to remove fragement to check for .json or .msgpack
+static std::string _GetPath(const std::string& uri)
+{
+    std::string path = uri;
+    size_t hashindex = path.find_last_of('#');
+
+    if (hashindex != std::string::npos) {
+        path = path.substr(0, hashindex);
+    }
+    return path;
+}
+
+/// \brief get the scheme of the uri, e.g. file: or openrave:
+static void ParseURI(const std::string& uri, std::string& scheme, std::string& path, std::string& fragment)
+{
+    path = uri;
+    size_t hashindex = path.find_last_of('#');
+    if (hashindex != std::string::npos) {
+        fragment = path.substr(hashindex + 1);
+        path = path.substr(0, hashindex);
+    }
+
+    size_t colonindex = path.find_first_of(':');
+    if (colonindex != std::string::npos) {
+        // notice: in python code, like realtimerobottask3.py, it pass scheme as {openravescene: mujin}. No colon,
+        scheme = path.substr(0, colonindex);
+        path = path.substr(colonindex + 1);
+    }
+
+}
+
+static std::string ResolveURI(const std::string& scheme, const std::string& path, const std::string& curdir, const std::vector<std::string>& vOpenRAVESchemeAliases)
+{
+    if (scheme.empty() && path.empty() ) {
+        return std::string();
+    }
+    else if (scheme == "file")
+    {
+        return RaveFindLocalFile(path, curdir);
+    }
+    else if (find(vOpenRAVESchemeAliases.begin(), vOpenRAVESchemeAliases.end(), scheme) != vOpenRAVESchemeAliases.end())
+    {
+        return RaveFindLocalFile(path, curdir);
+    }
+    return std::string();
+}
+
+/// \brief resolve a uri
+static std::string ResolveURI(const std::string& uri, const std::string& curdir, const std::vector<std::string>& vOpenRAVESchemeAliases)
+{
+    std::string scheme, path, fragment;
+    ParseURI(uri, scheme, path, fragment);
+
+    return ResolveURI(scheme, path, curdir, vOpenRAVESchemeAliases);
+}
+
+
+}
+
+
+
 class JSONRemoteHelper
 {
